+++ rich/traceback.py
@@ -42,6 +42,9 @@ from .theme import Theme
 
 WINDOWS = platform.system() == "Windows"
 
+LOCALS_MAX_LENGTH = 10
+LOCALS_MAX_STRING = 80
+
 
 def install(
     *,
@@ -146,8 +149,8 @@ class Traceback:
         show_locals (bool, optional): Enable display of local variables. Defaults to False.
         indent_guides (bool, optional): Enable indent guides in code and locals. Defaults to True.
         locals_max_length (int, optional): Maximum length of containers before abbreviating, or None for no abbreviation.
-            Defaults to None.
-        locals_max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to None.
+            Defaults to 10.
+        locals_max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to 80.
     """
 
     def __init__(
@@ -159,8 +162,8 @@ class Traceback:
         word_wrap: bool = False,
         show_locals: bool = False,
         indent_guides: bool = True,
-        locals_max_length: int = None,
-        locals_max_string: int = None,
+        locals_max_length: int = LOCALS_MAX_LENGTH,
+        locals_max_string: int = LOCALS_MAX_STRING,
     ):
         if trace is None:
             exc_type, exc_value, traceback = sys.exc_info()
@@ -193,8 +196,8 @@ class Traceback:
         word_wrap: bool = False,
         show_locals: bool = False,
         indent_guides: bool = True,
-        locals_max_length: int = None,
-        locals_max_string: int = None,
+        locals_max_length: int = LOCALS_MAX_LENGTH,
+        locals_max_string: int = LOCALS_MAX_STRING,
     ) -> "Traceback":
         """Create a traceback from exception info
 
@@ -209,8 +212,8 @@ class Traceback:
             show_locals (bool, optional): Enable display of local variables. Defaults to False.
             indent_guides (bool, optional): Enable indent guides in code and locals. Defaults to True.
             locals_max_length (int, optional): Maximum length of containers before abbreviating, or None for no abbreviation.
-                Defaults to None.
-            locals_max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to None.
+                Defaults to 10.
+            locals_max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to 80.
 
         Returns:
             Traceback: A Traceback instance that may be printed.
@@ -237,6 +240,8 @@ class Traceback:
         exc_value: BaseException,
         traceback: Optional[TracebackType],
         show_locals: bool = False,
+        locals_max_length: int = LOCALS_MAX_LENGTH,
+        locals_max_string: int = LOCALS_MAX_STRING,
     ) -> Trace:
         """Extract traceback information.
 
@@ -245,6 +250,9 @@ class Traceback:
             exc_value (BaseException): Exception value.
             traceback (TracebackType): Python Traceback object.
             show_locals (bool, optional): Enable display of local variables. Defaults to False.
+            locals_max_length (int, optional): Maximum length of containers before abbreviating, or None for no abbreviation.
+                Defaults to 10.
+            locals_max_string (int, optional): Maximum length of string before truncating, or None to disable. Defaults to 80.
 
         Returns:
             Trace: A Trace instance which you can use to construct a `Traceback`.
@@ -274,7 +282,11 @@ class Traceback:
                     lineno=line_no,
                     name=frame_summary.f_code.co_name,
                     locals={
-                        key: pretty.traverse(value)
+                        key: pretty.traverse(
+                            value,
+                            max_length=locals_max_length,
+                            max_string=locals_max_string,
+                        )
                         for key, value in frame_summary.f_locals.items()
                     }
                     if show_locals