+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added rich.abc.RichRenderable
 - Added expand_all to rich.pretty.install()
 - Added locals_max_length, and locals_max_string to Traceback and logging.RichHandler
+- Set defaults of max_length and max_string for Traceback to 10 and 80
 - Added disable argument to Progress
 
 ### Changed